+++ base/builder/template_project/starter.py
@@ -2,7 +2,11 @@
 # coding= utf-8
 
 from base.application.service import engage
+from base import __VERSION__
 
 if __name__ == '__main__':
-    engage(__file__)
+    if __VERSION__ > '1.0.1':
+        engage(__file__)
+    else:
+        engage()